ob_label class

ob_label class

Usage

ob_label(
  label = class_missing,
  p = class_missing,
  angle = class_missing,
  alpha = class_missing,
  color = class_missing,
  family = class_missing,
  fill = class_missing,
  fontface = class_missing,
  hjust = class_missing,
  label.color = class_missing,
  label.margin = ggplot2::margin(1, 1, 1, 1, "pt"),
  label.padding = ggplot2::margin(2, 2, 2, 2, "pt"),
  label.r = class_missing,
  label.size = class_missing,
  lineheight = class_missing,
  polar_just = class_missing,
  nudge_x = class_missing,
  nudge_y = class_missing,
  size = class_missing,
  straight = class_missing,
  text.color = class_missing,
  vjust = class_missing,
  style = class_missing,
  plot_point = FALSE,
  position = 0.5,
  spacing = class_missing,
  x = class_missing,
  y = class_missing,
  ...
)

Arguments

label

text label

p

ob_point

angle

angle of text

alpha

numeric value for alpha transparency

color

character string for color

family

font family

fill

character string for fill color

fontface

Can be plain, bold, italic, or bold.italic

hjust

horizontal justification.

label.color

Color of label outline.

label.margin

Amount of distance around label. Unit vector of length four. Usually created with ggplot2::margin.

label.padding

Amount of padding around label. Unit vector of length four. Usually created with ggplot2::margin.

label.r

Radius of rounded corners. Defaults to 0.15 lines.

label.size

Width of label outline.

lineheight

Height of line of text

polar_just

an angle, polar point, or point that alters hjust and vjust (polar polar_just not stored in style)

nudge_x

Horizontal adjustment to nudge labels by.

nudge_y

Vertical adjustment to nudge labels by.

size

numeric size

straight

logical. If TRUE, make bzpath label text straight instead of curved.

text.color

Color of label text.

vjust

vertical justification

style

a style list

plot_point

plot ob_point (default = FALSE)

position

position (used in conjunction with the place function)

spacing

letter spacing for labels used with ob_path and ob_bezier

x

x-coordinate of center point. If specified, overrides x-coordinate of @p.

y

x-coordinate of center point. If specified, overrides y-coordinate of @p.

...

<dynamic-dots> properties passed to style