create double-headed arrow paths indicating variance

create double-headed arrow paths indicating variance

Usage

ob_variance(
  x,
  where = "north",
  theta = 50,
  bend = 0,
  looseness = 1,
  arrow_head = arrowheadr::arrow_head_deltoid(d = 2.3, n = 100),
  resect = 2,
  ...
)

Arguments

x

object

where

angle or named direction (e.g.,northwest, east, below, left)

theta

angle width

bend

Angle by which the control points are rotated

looseness

distance of control points as a ratio of the distance to the object's center (e.g., in a circle of radius 1, looseness = 1.5 means that that the control points will be 1.5 units from the start and end points.)

arrow_head

A 2-column matrix of polygon points

resect

A numeric(1) denoting millimeters or to shorten the arrow head and fins.

...

<dynamic-dots> properties passed to style