The pgon (polygon) class

A polygon is specified with a point object that contains at least 3 points, the start and the end. Any number of intermediate points are possible. The pgon class could not be named 'polygon' because that function name is used in base R graphics.

Usage

pgon(
  p = class_missing,
  label = class_missing,
  alpha = class_missing,
  color = class_missing,
  fill = class_missing,
  linewidth = 0.75,
  linetype = class_missing,
  style = class_missing,
  ...
)

Arguments

p

point object or list of point objects

label

A character, angle, or label object

alpha

numeric value for alpha transparency

color

character string for color

fill

character string for fill color

linewidth

Width of lines

linetype

type of lines

style

Gets and sets the styles associated with polygons

...

<dynamic-dots> properties passed to style

Details

If you wish to specify multiple polygons, you must supply a list of point objects. When plotted, the pgon function uses the ggplot2::geom_polygon function to create the geom.

Slots

length

The number of polygons in the pgon object

tibble

Gets a tibble (data.frame) containing parameters and styles used by ggplot2::geom_polygon.