Find projection of a point on an object (e.g., line or segment)

Find projection of a point on an object (e.g., line or segment)

Usage

projection(p, object, ...)

Arguments

p

ob_point

object

object (e.g., line or segment)

...

<dynamic-dots> arguments passed to style object