Make a variant of a function with alternate defaults

Makes a copy of a function with new defaults. Similar to purrr::partial except that arguments with new defaults still accept input.

Usage

redefault(.f, ...)

Arguments

.f

function

...

<dynamic-dots> new defaults

Value

function

Examples

squircle <- redefault(ob_ellipse, m1 = 4)
squircle(a = 3)
#> S3<ggdiagram::ob_ellipse/ggdiagram::centerpoint/ggdiagram::xy/ggdiagram::shape/ggdiagram::has_style/S7_object>
#> @ center:S3<ggdiagram::ob_point/ggdiagram::xy/ggdiagram::shape/ggdiagram::has_style/S7_object>
#> @ x: num 0
#> @ y: num 0
#> Other props: alpha, color, fill, shape, size, stroke, auto_label,
#>              bounding_box, centroid, length, r, theta, style,
#>              tibble, xy, geom, label, place, aesthetics
#> @ a     : num 3
#> @ b     : num 3
#> @ angle :S3<ggdiagram::radian/ggdiagram::ob_angle/double/S7_object>
#> @ radian: num 0
#> 
#> @ m1    : num 4
#> @ m2    : num 4
#> Other props: label, alpha, color, fill, linewidth, linetype,
#>              n, area, bounding_box, focus_1, focus_2, length,
#>              perimeter, polygon, style, tibble, angle_at, geom,
#>              normal_at, place, point_at, tangent_at, aesthetics