ob_circle class

ob_circle class

Usage

ob_circle(
  center = ob_point(0, 0),
  radius = 1,
  label = character(0),
  alpha = numeric(0),
  color = character(0),
  fill = character(0),
  linewidth = numeric(0),
  linetype = numeric(0),
  n = numeric(0),
  style = S7::class_missing,
  x0 = numeric(0),
  y0 = numeric(0),
  ...
)

Arguments

center

point at center of the circle

radius

distance between center and edge circle

label

A character, angle, or label object

alpha

numeric value for alpha transparency

color

character string for color

fill

character string for fill color

linewidth

Width of lines

linetype

type of lines

n

number of points in circle (default = 360)

style

an ob_style object

x0

x-coordinate of center point. If specified, overrides x-coordinate of @center.

y0

x-coordinate of center point. If specified, overrides y-coordinate of @center.

...

<dynamic-dots> arguments passed to style object

Value

ob_circle object

Slots

aesthetics

A list of information about the circle's aesthetic properties

angle_at

A function that finds the angle of the specified point in relation to the circle's center

area

area of the circle

bounding_box

a rectangle that contains all the circles

circumference

circumference of the circle

geom

A function that converts the object to a geom. Any additional parameters are passed to ggforce::geom_circle.

length

The number of circles in the circle object

normal_at

A function that finds a point that is perpendicular from the circle and at a specified distance

point_at

A function that finds a point on the circle at the specified angle.

polygon

a tibble containing information to create all the polygon points in a circle.

tangent_at

A function that finds the tangent line at the specified angle.

tibble

Gets a tibble (data.frame) containing parameters and styles used by ggforce::geom_cirlce.

Examples

# specify center point and radius
p <- ob_point(0,0)
ob_circle(p, radius = 6)
#> S3<ggdiagram::ob_circle/ggdiagram::centerpoint/ggdiagram::xy/ggdiagram::shape/ggdiagram::has_style/S7_object>
#> @ center:S3<ggdiagram::ob_point/ggdiagram::xy/ggdiagram::shape/ggdiagram::has_style/S7_object>
#> @ x: num 0
#> @ y: num 0
#> Other props: alpha, color, fill, shape, size, stroke, auto_label,
#>              bounding_box, centroid, length, r, theta, style,
#>              tibble, xy, geom, label, place, aesthetics
#> @ radius: num 6
#> Other props: label, alpha, color, fill, linewidth, linetype,
#>              n, area, bounding_box, circumference, diameter,
#>              length, polygon, style, tibble, geom, arc, angle_at,
#>              normal_at, tangent_at, place, point_at, aesthetics